refactor: replace <script> tag with es6 module when using api functions
