+++ static/js/api/apiParameters.js
@@ -1,0 +1,16 @@
+let apiModel = {
+  route_attraction: "/attraction/",
+  requestParameters: {
+    cache: "no-cache",
+    credentials: "same-origin",
+    headers: {
+      "user-agent": "Mozilla/4.0 MDN Example",
+      "content-type": "application/json",
+    },
+    mode: "cors",
+    redirect: "follow",
+    referrer: "no-referrer",
+  },
+};
+
+export { apiModel };
