+++ templates/attraction.html
@@ -9,8 +9,7 @@
     />
     <script src="../static/js/ui/common.js"></script>
     <script src="../static/js/ui/member.js"></script>
-    <script src="../static/js/ui/APIs.js"></script>
-    <script src="../static/js/attraction.js"></script>
+    <script type="module" src="../static/js/attraction.js"></script>
   </head>
   <body>
     <script src="../static/js/components/navbar.js"></script>
