+++ templates/booking.html
@@ -5,8 +5,7 @@
     <link rel="stylesheet" type="text/css" href="../static/css/booking.css" />
     <script src="../static/js/ui/common.js"></script>
     <script src="../static/js/ui/member.js"></script>
-    <script src="../static/js/ui/APIs.js"></script>
-    <script src="../static/js/booking.js"></script>
+    <script type="module" src="../static/js/booking.js"></script>
   </head>
   <body>
     <script src="../static/js/components/navbar.js"></script>
