+++ templates/index.html
@@ -4,8 +4,7 @@
     <link rel="stylesheet" type="text/css" href="../static/css/dialog.css" />
     <script src="../static/js/ui/common.js"></script>
     <script src="../static/js/ui/member.js"></script>
-    <script src="../static/js/ui/APIs.js"></script>
-    <script src="../static/js/index.js"></script>
+    <script type="module" src="../static/js/index.js"></script>
   </head>
 
   <body>
